Stop, Restart, Terminate on runtime, also from scripts, pluse shortcuts Ctrl+C, Ctrl+R, Ctrl+X

+++ model/src/com/bigcustard/blurp/model/Timer.java
@@ -9,9 +9,9 @@ public abstract class Timer {
         stopwatch = new Stopwatch();
     }
 
-    public abstract Timer after(long milliseconds, Runnable action);
+    public abstract Timer after(int milliseconds, Runnable action);
 
-    public abstract Timer every(long milliseconds, Runnable action);
+    public abstract Timer every(int milliseconds, Runnable action);
 
     public abstract Timer remove(Runnable action);
